Refactor: Generic SliceViewChunk and SliceViewChunkSource
By refactoring the VolumeChunkSource out of the SliceView code, we can create multiple spatial chunk sources similar to the existing VolumeChunkSource, but containing different data.
