+++ src/neuroglancer/datasource/ndstore/backend.ts
@@ -16,7 +16,7 @@
 
 import {registerChunkSource} from 'neuroglancer/chunk_manager/backend';
 import {VolumeChunkSourceParameters} from 'neuroglancer/datasource/ndstore/base';
-import {ParameterizedVolumeChunkSource, VolumeChunk} from 'neuroglancer/sliceview/backend';
+import {ParameterizedVolumeChunkSource, VolumeChunk} from 'neuroglancer/sliceview/volume/backend';
 import {ChunkDecoder} from 'neuroglancer/sliceview/backend_chunk_decoders';
 import {decodeJpegChunk} from 'neuroglancer/sliceview/backend_chunk_decoders/jpeg';
 import {decodeNdstoreNpzChunk} from 'neuroglancer/sliceview/backend_chunk_decoders/ndstoreNpz';
