+++ src/neuroglancer/datasource/nifti/base.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {DataType, VolumeType} from 'neuroglancer/sliceview/base';
+import {DataType, VolumeType} from 'neuroglancer/sliceview/volume/base';
 import {mat4, quat, vec3} from 'neuroglancer/util/geom';
 
 export const GET_NIFTI_VOLUME_INFO_RPC_ID = 'nifti/getNiftiVolumeInfo';
