+++ src/neuroglancer/sliceview/backend_chunk_decoders/jpeg.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {VolumeChunk} from 'neuroglancer/sliceview/backend';
+import {VolumeChunk} from 'neuroglancer/sliceview/volume/backend';
 import {postProcessRawData} from 'neuroglancer/sliceview/backend_chunk_decoders/postprocess';
 import {decodeJpegStack} from 'neuroglancer/sliceview/decode_jpeg_stack';
