+++ src/neuroglancer/sliceview/single_texture_chunk_format.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import {ChunkFormat, VolumeChunk, VolumeChunkSource} from 'neuroglancer/sliceview/frontend';
+import {VolumeChunk, VolumeChunkSource} from 'neuroglancer/sliceview/volume/frontend';
+import {ChunkFormat} from 'neuroglancer/sliceview/frontend';
 import {TypedArray} from 'neuroglancer/util/array';
 import {Disposable, RefCounted} from 'neuroglancer/util/disposable';
 import {GL} from 'neuroglancer/webgl/context';
