+++ src/neuroglancer/sliceview/uncompressed_chunk_format.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-import {DataType, VolumeChunkSpecification} from 'neuroglancer/sliceview/base';
-import {ChunkFormatHandler, registerChunkFormatHandler, VolumeChunkSource} from 'neuroglancer/sliceview/frontend';
+import {DataType, VolumeChunkSpecification} from 'neuroglancer/sliceview/volume/base';
+import {VolumeChunkSource} from 'neuroglancer/sliceview/volume/frontend';
+import {ChunkFormatHandler, registerChunkFormatHandler} from 'neuroglancer/sliceview/frontend';
 import {SingleTextureChunkFormat, SingleTextureVolumeChunk} from 'neuroglancer/sliceview/single_texture_chunk_format';
 import {TypedArray, TypedArrayConstructor} from 'neuroglancer/util/array';
 import {RefCounted} from 'neuroglancer/util/disposable';
